// apps/console/src/__generated__/core/ThirdPartyThirdPartiesPageFragment.graphql.ts
export type BusinessImpact = "CRITICAL" | "HIGH" | "LOW" | "MEDIUM";
export type DataSensitivity = "CRITICAL" | "HIGH" | "LOW" | "MEDIUM" | "NONE";
import { FragmentRefs } from "relay-runtime";
export type ThirdPartyThirdPartiesPageFragment$data = {
  readonly childThirdParties: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly name: string;
        readonly riskAssessments: {
          readonly edges: ReadonlyArray<{
            readonly node: {
              readonly businessImpact: BusinessImpact;
              readonly createdAt: string;
              readonly dataSensitivity: DataSensitivity;
              readonly id: string;
            };
          }>;
        };
        readonly websiteUrl: string | null | undefined;
      };
    }>;
  };
  readonly id: string;
  readonly " $fragmentType": "ThirdPartyThirdPartiesPageFragment";
};
export type ThirdPartyThirdPartiesPageFragment$key = {
  readonly " $data"?: ThirdPartyThirdPartiesPageFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"ThirdPartyThirdPartiesPageFragment">;
};